python/subscribe/requirements.txt
remotivelabs-broker ~= 0.9.1
